API

Interface

PlanerTransformer.PCB_Specification — Type.

PCB_Specification(trace_edge_gap,
                  trace_trace_gap,
                  outer_copper_thickness,
                  inner_copper_thickness = 0.0,
                  number_of_layers = 2,
                  ρ_20=1.68e-8,
                  temperature_coefficient=0.003862)

Store PCB data.

** Fields **

trace_edge_gap – minimum distance from trace to PCB edge (m)

trace_trace_gap – minimum distance between traces (m)

outer_copper_thickness – thickness of top and bottom copper layers (m)

inner_copper_thickness – thickness of inner copper layers (m)

number_of_layers – number of copper layers

ρ_20 – conductivity, default to 1.68e-8 Ωm for Cu @ 20C

temperature_coefficient – default to 0.003862 1/K for Cu

source

PlanerTransformer.copper_weight_to_meters — Function.

copper_weight_to_meters(oz)

PCB copper thickness is typicaly given in ounces. This function multiplies by 0.48e-3 to give thickness in meters.

source

PlanerTransformer.CoreGeometry — Type.

CoreGeometry

Fields

winding_aperture – m

half_center_width – m

center_length – m

effective_volume – m^3

effective_area – m^2

effective_length – m

mass – Kg

source

PlanerTransformer.FerriteProperties — Type.

FerriteProperties(frequency_range, troom, thot, bh_room, bh_hot, spl_hot)

Store material data.

Fields

fmin – minimum recommended operating frequency

fmax – maximum recommended operating frequency

troom – typicaly 25C

thot – typicaly 100C

bh_room – BH loop at room temperature

bh_hot – BH loop at hot temperature

spl_hot – specfic power loss data at hot temperature

source

PlanerTransformer.BHloop — Type.

BHloop(hc,bs,br)

Define the BH loop of a magnetic material.

Fields

hc – Coercive force (A-turn/m)

bs – Remnant flux density (Tesla)

br – Saturation flux density (Tesla)

source

PlanerTransformer.SpecificPowerLossData — Type.

SpecificPowerLossData(frequency::Tuple, mb::Tuple)
SpecificPowerLossData(input::SplInput)

Capture specific power loss data from datasheet.

Data is stored as a series of linear approximations on a log log plot, one for each frequency.

Fields

frequency – frequency of linear approximation (Hz)

mb – Tuple (slope, offset) defining the linear approximation

In order to simplify manual data entry, data my also be passed as a SplInput object. All data is in MKS units (Hz, Tesla, W/m^3).

source

PlanerTransformer.WindingLayer — Type.

WindingLayer

Fields

width – width of trace (m)

length – length of trace (m)

thickness – thickness of trace (m)

number_of_turns – number of turns

source

PlanerTransformer.Winding — Type.

Winding(pcb::PCB_Specification,
        windinglayers,
        isseries::Bool)

Create a Winding by combining several WindingLayers.

source

PlanerTransformer.Magnetics — Type.

Magnetics(fp::FerriteProperties, cores::Array{CoreGeometry,1})

All magnetic information for Transformer in one object.

source

PlanerTransformer.Transformer — Type.

Transformer(m::Magnetics,w::Array{Winding,1})

Transformer is a combination of Magnetics and two or more windings.

source

PlanerTransformer.turns — Function.

turns(wl::WindingLayer)
turns(w::Winding)
turns(t::Transformer)

Number of turns, Array for Transformer.

source

PlanerTransformer.winding_resistance — Function.

winding_resistance(wl::WindingLayer, ρ)
winding_resistance(wl::WindingLayer, pcb::PCB_Specification, temperature=100.0)
winding_resistance(w::Winding, temperature=100.0)

Returns the resistance of a Winding or WindingLayer.

source

PlanerTransformer.TransformerPowerDissipation — Type.

TransformerPowerDissipation(t::Transformer, input::Array{Float64,1}, frequency)

Computes power dissipation of transformer.

The first element of the input array is the peak to peak voltage applied to the first winding. The following elements are the load currents in the output windings. Returns a TransformerPowerDissipation object.

Fields

transformer – from input

frequency – from input

flux_density – peak flux density (Tesla)

winding_voltage – peak to peak voltage on each winding

core_specific_power – power dissipated in core (W/m^3)

core_total_power – power dissipated in core (W)

winding_power – power dissipated in each winding (W)

total_power – core_total_power +sum(winding_power)

source

PlanerTransformer.volt_seconds_per_turn — Function.

volt_seconds_per_turn(effective_area, flux_density_pp)
volt_seconds_per_turn(cg::CoreGeometry, flux_density_pp)
volt_seconds_per_turn(m::Magnetics, flux_density_pp)
volt_seconds_per_turn(t::Transformer, flux_density_pp)

Volt seconds per turn at flux_density_pp. The first form is just an alias for multiply.

source

PlanerTransformer.volt_seconds — Function.

volt_seconds(t::Transformer, flux_density_pp)

Volt seconds at flux_density_pp.

source

PlanerTransformer.volts_per_turn — Function.

volts_per_turn(cg::CoreGeometry,
               fp::FerriteProperties, loss_limit, frequency)

Returns the maximum volts per turn.

source

PlanerTransformer.specificpowerloss — Function.

specificpowerloss(spl::SpecificPowerLossData, flux_density, frequency)
specificpowerloss(fp::FerriteProperties, flux_density, frequency)
specificpowerloss(m::Magnetics, flux_density, frequency)
specificpowerloss(t::Transformer, flux_density, frequency)

Returns specific power loss.

source

PlanerTransformer.flux_density — Function.

flux_density(spl::SpecificPowerLossData, coreloss, frequency)
flux_density(fp::FerriteProperties, coreloss, frequency)
flux_density(m::Magnetics, coreloss, frequency)
flux_density(t::Transformer, coreloss, frequency)

Returns magnetic field strength in Tesla.

source

PlanerTransformer.SplInput — Type.

SplInput(data::Tuple)

Simplify manual input of specific power loss data.

Data format is as follows. (f1,(x1,y1),(x2,y2), f2,(x3,y3),(x4,y4), ...)

Elements of Tuple

fn – frequency (Hz)

xn – flux density (mT)

yn – specific power loss (Kw/m^3)

The only purpose of this object is to pass to SpecificPowerLossData. Data is converted to MKS units there.

source