Data Entry API

Data Entry API

Use these if core_geometry_dict and ferrite_dict do not have the parts you need.

PlanerTransformer.CoreGeometry — Type.

CoreGeometry

Fields

name – string identifying the core winding_aperture_height – m

winding_aperture – m

half_center_width – m

center_length – m

effective_volume – m^3

effective_area – m^2

effective_length – m

mass – Kg

source

PlanerTransformer.FerriteProperties — Type.

FerriteProperties(frequency_range, troom, thot, bh_room, bh_hot, spl_hot)

Store material data.

Fields

name – string identifying the ferrite

fmin – minimum recommended operating frequency

fmax – maximum recommended operating frequency

troom – typicaly 25C

thot – typicaly 100C

bh_room – BH loop at room temperature

bh_hot – BH loop at hot temperature

spl_hot – specfic power loss data at hot temperature

source

PlanerTransformer.BHloop — Type.

BHloop(hc,bs,br)

Define the BH loop of a magnetic material.

Fields

hc – Coercive force (A-turn/m)

bs – Remnant flux density (Tesla)

br – Saturation flux density (Tesla)

source

PlanerTransformer.SpecificPowerLossData — Type.

SpecificPowerLossData(frequency::Tuple, mb::Tuple)
SpecificPowerLossData(input::SplInput)

Capture specific power loss data from datasheet.

Data is stored as a series of linear approximations on a log log plot, one for each frequency.

Fields

frequency – frequency of linear approximation (Hz)

mb – Tuple (slope, offset) defining the linear approximation

In order to simplify manual data entry, data my also be passed as a SplInput object. All data is in MKS units (Hz, Tesla, W/m^3).

source

PlanerTransformer.SplInput — Type.

SplInput(data::Tuple)

Simplify manual input of specific power loss data.

Data format is as follows. (f1,(x1,y1),(x2,y2), f2,(x3,y3),(x4,y4), ...)

Elements of Tuple

fn – frequency (Hz)

xn – flux density (mT)

yn – specific power loss (Kw/m^3)

The only purpose of this object is to pass to SpecificPowerLossData. Data is converted to MKS units there.

source

PlanerTransformer.Magnetics_ER_Input — Type.

Magnetics_ER_Input

Direct entry of Magnetics ER core values from datasheet.

Fields

name –

effective_length – mm

effective_area – mm^2

effective_volume – mm^3

mass – g

e – mm

f – mm

aperture height – mm

source