Internal API

Internal

PlanerTransformer.center_frequency — Function.

center_frequency(x :: FerriteProperties)
center_frequency(x :: Transformer)

Middle of the minimum and maximum recomended operating frequencys from datasheet.

source

PlanerTransformer.winding_breadth_volume — Function.

winding_breadth_volume(windings)

Tuple (breadth, volume). For internal use calculating leakage inductance.

source

PlanerTransformer.winding_geometry — Function.

winding_geometry(pcb, core, turns)

Create a WindingGeometry object.

source

PlanerTransformer.winding_area — Function.

winding_area(x)

Area the windings occupy. For internal use.

source

PlanerTransformer.power_error — Function.

power_error(ta::TransformerPowerAnalysis)

returns the difference between total_power and the power loss computed from the voltages and currents.

source

PlanerTransformer.pcb_thickness — Function.

pcb_thickness(x)

Overall thickness of PCB.

source

PlanerTransformer.sides — Function.

sides(isprimary,index)

For a given layer (index), returns the number (1 or 2) of sides which are part of the same winding.

This function is for internal use calculating effectiveresistance and leakageinductance.

source