Design API

Design

Note

All fields of structs can be accessed as functions.

PCB

PlanerTransformer.Stackup — Type.

Stackup(material,thickness)

Specify the PCB stackup.

example: for double sided FR408 PCB with 1oz copper 63mil thick. Stackup([copper,fr408,copper],[0.48e-3,1.6e-3 - (2*0.48e-3),0.48e-3])

copper and fr408 are exported constants.

source

PlanerTransformer.LayerMaterial — Type.

LayerMaterial

Subtypes of LayerMaterial hold the material properties for the layers of the PCB.

source

PlanerTransformer.Dielectric — Type.

Dielectric(name, ϵ)

Fields

name – name of the material

ϵ – permittivity

source

PlanerTransformer.Conductor — Type.

Conductor(name, ρ, tc)

Fields

name – name of the material

ρ_20 – conductivity at 20C

tc – temperature coefficient

see constant copper as example.

source

PlanerTransformer.copper_weight_to_meters — Function.

copper_weight_to_meters(oz)

PCB copper thickness is typicaly given in ounces. This function multiplies by 35e-6 to give thickness in meters.

source

PlanerTransformer.PCB_Specification — Type.

PCB_Specification(trace_edge_gap  :: Float64,
                  trace_trace_gap :: Float64,
                  stackup         :: Stackup)

Store PCB data.

Fields

trace_edge_gap – minimum distance from trace to PCB edge (m)

trace_trace_gap – minimum distance between traces (m)

stackup – defines material and thickness of the layers

source

Windings

PlanerTransformer.Windings — Type.

Windings

Fields

pcb – PCB_Specification

core – CoreGeometry

primarywindinggeometry – WindingGeometry for all layers of primary

secondarywindinggeometry– WindingGeometry for all layers of secondary

isprimary – tuple of Bool for each conductor layer

sides – for internal use

isprimaryseries – true if primary is connected in series

issecondaryseries – true if secondary is connected in series

Create Windings objects with windings.

source

PlanerTransformer.windings — Function.

windings(pcb, core,
        primaryturnsperlayer, secondaryturnsperlayer,
        isprimary,
        isprimaryeries, issecondaryseries)

Create a Windings object.

source

PlanerTransformer.WindingGeometry — Type.

WindingGeometry

2D representation of one layer of a winding.

Fields

width – width of trace (m)

length – length of trace (m)

turns – number of turns

WindingGeometry objects are typicaly created with winding_geometry

source

PlanerTransformer.winding_resistance — Function.

winding_resistance(x :: Windings, frequency=0.0, temperature=100.0)
winding_resistance(x :: Transformer, frequency=0.0, temperature=100.0)
winding_resistance(x :: TransformerPowerAnalysis, frequency=0.0)

Returns tuple (primary_resistance, secondary_resistance)

source

PlanerTransformer.turns — Function.

turns(x :: WindingGeometry)
turns(x :: Windings)
turns(x :: Transformer)
turns(x :: TransformerPowerAnalysis)

Number of turns, Tuple for objects with two windings.

source

PlanerTransformer.capacitance — Function.

capacitance(x)

Capacitance between primary and secondary.

source

PlanerTransformer.leakage_inductance — Function.

leakage_inductance(x)

Leakage inductance from volume between primary and secondary, referenced to primary. Fringing fields are not taken into account.

source

PlanerTransformer.does_pcb_fit — Function.

does_pcb_fit(t)

true if PCB thickness is <= winding aperature height.

source

Transformer

PlanerTransformer.Transformer — Type.

Transformer(ferrite, windings)

Fields

ferrite – FerriteProperties object

windings – Windings object

source

PlanerTransformer.volt_seconds_per_turn — Function.

volt_seconds_per_turn(effective_area, flux_density_pp)
volt_seconds_per_turn(x ::CoreGeometry, flux_density_pp)
volt_seconds_per_turn(x ::Windings, flux_density_pp)
volt_seconds_per_turn(x ::Transformer, flux_density_pp)
volt_seconds_per_turn(x ::TransformerPowerAnalysis, flux_density_pp)
volt_seconds_per_turn(x ::TransformerPowerAnalysis)

Volt seconds per turn at flux_density_pp. The first form is just an alias for multiply.

source

PlanerTransformer.volt_seconds — Function.

volt_seconds(x :: Windings, flux_density_pp)
volt_seconds(x :: Transformer, flux_density_pp)
volt_seconds(x :: TransformerPowerAnalysis, flux_density_pp)
volt_seconds(x :: TransformerPowerAnalysis)

Volt seconds at flux_density_pp.

source

PlanerTransformer.flux_density — Function.

flux_density(spl::SpecificPowerLossData, coreloss, frequency)
flux_density(fp::FerriteProperties, coreloss, frequency)
flux_density(t::Transformer, coreloss, frequency)
flux_density(tpa::TransformerPowerAnalysis)

Returns magnetic field strength in Tesla.

source

PlanerTransformer.specific_power_loss — Function.

specific_power_loss(spl::SpecificPowerLossData, flux_density, frequency)
specific_power_loss(fp::FerriteProperties, flux_density, frequency)
specific_power_loss(t::Transformer, flux_density, frequency)
specific_power_loss(tpa::TransformerPowerAnalysis)

Returns specific power loss.

source

Analysis

PlanerTransformer.TransformerPowerAnalysis — Type.

TransformerPowerAnalysis

Object to store the result of a transformerpoweranalysis().

Fields

transformer – from input

converter – type of converter.

flux_density – peak flux density in Tesla

voltage – (primary, secondary) Vpp

current – (promary, secondary) Ipp

ac_winding_resistance – (primary, secondary) Ω

dc_winding_resistance – (primary, secondary) Ω

equilivent_resistance – (primary, secondary) Ω

v_core – voltage, after drop accross input resistance, referenced to 1 turn

core_specific_power – power dissipated in core in W/m^3

core_total_power – power dissipated in core in W

source

PlanerTransformer.transformer_power_analysis — Function.

TransformerPowerAnalysis(t::Transformer,
                         c::Converter)

Determine the losses of a transformer.

source

PlanerTransformer.winding_power — Function.

winding_power(tpa::TransformerPowerAnalysis)

Power dissipation due to windign resistance. Tuple (primary, secondary).

source

PlanerTransformer.total_power — Function.

total_power(tpa::TransformerPowerAnalysis)

Total power dissipation of transformer. Sum of winding and core losses.

source

PlanerTransformer.input_power — Function.

input_power(tpa::TransformerPowerAnalysis)

Average power delivered to the primary of the transformer.

source

PlanerTransformer.output_power — Function.

output_power(tpa::TransformerPowerAnalysis)

Average power from the secondary of the transformer.

source

PlanerTransformer.efficiency — Function.

efficiency(tpa::TransformerPowerAnalysis)

Efficiency (-Pout/Pin) of transformer.

source

PlanerTransformer.r_core — Function.

r_core(tpa::TransformerPowerAnalysis)

Resistance, referenced to 1 turn, for spice model to match core power dissipation.

source

PlanerTransformer.chan_inductor — Function.

chan_inductor(ferriteproperties, effective_area, effective_length,ishot=true)
chan_inductor(transformer, ishot=true)
chan_inductor(transformer_power_analysis, ishot=true)

Parameters for LTspice Chan inductor to be used for magnetizing inductance.

source