Design API

Design

PlanerTransformer.PCB_Specification — Type.

PCB_Specification(trace_edge_gap,
                  trace_trace_gap,
                  outer_copper_thickness,
                  inner_copper_thickness = 0.0,
                  number_of_layers = 2,
                  ρ_20=1.68e-8,
                  temperature_coefficient=0.003862)

Store PCB data.

** Fields **

trace_edge_gap – minimum distance from trace to PCB edge (m)

trace_trace_gap – minimum distance between traces (m)

outer_copper_thickness – thickness of top and bottom copper layers (m)

inner_copper_thickness – thickness of inner copper layers (m)

number_of_layers – number of copper layers

ρ_20 – conductivity, default to 1.68e-8 Ωm for Cu @ 20C

temperature_coefficient – default to 0.003862 1/K for Cu

source

PlanerTransformer.copper_weight_to_meters — Function.

copper_weight_to_meters(oz)

PCB copper thickness is typicaly given in ounces. This function multiplies by 0.48e-3 to give thickness in meters.

source

PlanerTransformer.WindingLayer — Type.

WindingLayer

Fields

width – width of trace (m)

length – length of trace (m)

thickness – thickness of trace (m)

number_of_turns – number of turns

source

PlanerTransformer.winding_layer — Function.

winding_layer(pcb :: PCB_Specification,
             isouter :: Bool,
             core :: CoreGeometry,
             number_of_turns :: Int)

Create a WindingLayer for a specific core and PCB.

source

PlanerTransformer.Winding — Type.

Winding(pcb::PCB_Specification,
        windinglayers,
        isseries::Bool)

Create a Winding by combining several WindingLayers.

source

PlanerTransformer.Magnetics — Type.

Magnetics(fp::FerriteProperties, cores::Array{CoreGeometry,1})

All magnetic information for Transformer in one object.

source

PlanerTransformer.Transformer — Type.

Transformer(m::Magnetics,w::Array{Winding,1})

Transformer is a combination of Magnetics and two or more Windings.

source

PlanerTransformer.turns — Function.

turns(windinglayer::WindingLayer)
turns(winding::Winding)
turns(transformer::Transformer)

Number of turns, Array for Transformer.

source

PlanerTransformer.winding_resistance — Function.

winding_resistance(wl::WindingLayer, ρ)
winding_resistance(wl::WindingLayer, pcb::PCB_Specification, temperature=100.0)
winding_resistance(w::Winding, temperature=100.0)

Returns the resistance of a Winding or WindingLayer.

source

PlanerTransformer.TransformerPowerDissipation — Type.

TransformerPowerDissipation(t::Transformer, input::Array{Float64,1}, frequency)

Computes power dissipation of transformer.

The first element of the input array is the peak to peak voltage applied to the first winding. The following elements are the load currents in the output windings. Returns a TransformerPowerDissipation object.

Fields

transformer – from input

frequency – from input

flux_density – peak flux density (Tesla)

winding_voltage – peak to peak voltage on each winding

core_specific_power – power dissipated in core (W/m^3)

core_total_power – power dissipated in core (W)

winding_power – power dissipated in each winding (W)

total_power – core_total_power +sum(winding_power)

source

PlanerTransformer.volt_seconds_per_turn — Function.

volt_seconds_per_turn(effective_area, flux_density_pp)
volt_seconds_per_turn(cg::CoreGeometry, flux_density_pp)
volt_seconds_per_turn(m::Magnetics, flux_density_pp)
volt_seconds_per_turn(t::Transformer, flux_density_pp)

Volt seconds per turn at flux_density_pp. The first form is just an alias for multiply.

source

PlanerTransformer.volt_seconds — Function.

volt_seconds(t::Transformer, flux_density_pp)

Volt seconds at flux_density_pp.

source

PlanerTransformer.equivalent_parallel_resistance — Function.

equivalent_parallel_resistance(tansformer::Transformer,
                               volts,
                               frequency=center_frequency(transformer),
                               turns=1,
                               ishot::Bool=true)

Parallel resistance for spice model for correct power dissipation.

source

PlanerTransformer.chan_inductor — Function.

chan_inductor(ferriteproperties, effective_area, effective_length,ishot=true)
chan_inductor(magnetics, ishot=true)
chan_inductor(transformer, ishot=true)

Parameters for LTspice Chan inductor to be used for magnetizing inductance.

source

PlanerTransformer.volts_per_turn — Function.

volts_per_turn(cg::CoreGeometry,
               fp::FerriteProperties, loss_limit, frequency)

Returns the maximum volts per turn.

source

PlanerTransformer.specific_power_loss — Function.

specific_power_loss(spl::SpecificPowerLossData, flux_density, frequency)
specific_power_loss(fp::FerriteProperties, flux_density, frequency)
specific_power_loss(m::Magnetics, flux_density, frequency)
specific_power_loss(t::Transformer, flux_density, frequency)

Returns specific power loss.

source

PlanerTransformer.flux_density — Function.

flux_density(spl::SpecificPowerLossData, coreloss, frequency)
flux_density(fp::FerriteProperties, coreloss, frequency)
flux_density(m::Magnetics, coreloss, frequency)
flux_density(t::Transformer, coreloss, frequency)

Returns magnetic field strength in Tesla.

source